Allow deletion of editor windows
When they're deleted we remove them from the CodeModels collection and
ensure that gridster removes the corresponding html -- we do not want
them removed directly via a simple @$el.remove() (or gridster will
freak out).

To pull this off we remove elements via the onBeforeClose hook, telling
gridster to call remove_widget on the $el. Doing this results in an
$el.remove() being called *after* remove_widget has already been
performed, but it doesn't seem to cause any problems.