since we got rid of bootstrap in previous commit, we need more css
Changes to css include floating the main and sidebar regions left, as
well as some other changes.

One possibly stupid change was defining the global styles for ul and
li to get rid of padding, this is probably silly and we should scope
it instead -- scoping should probably be under the 'repl' section as i
think it was the padding/margin to the 'li' elements for each repl
that was causing the alignment issues.